requirements.txt
openai>=2.20.0
python-dotenv>=1.2.1
